DACT-531 - removing unneeded duplicate code.
