+++ src/main/java/uk/gov/companieshouse/officerfiling/api/validation/OfficerAppointmentValidator.java
@@ -7,19 +7,12 @@ import java.util.Optional;
 import javax.servlet.http.HttpServletRequest;
 import uk.gov.companieshouse.api.error.ApiError;
 import uk.gov.companieshouse.api.model.company.CompanyProfileApi;
-import uk.gov.companieshouse.api.model.delta.officers.AppointmentFullRecordAPI;
 import uk.gov.companieshouse.api.model.transaction.Transaction;
 import uk.gov.companieshouse.logging.Logger;
 import uk.gov.companieshouse.officerfiling.api.enumerations.ApiEnumerations;
 import uk.gov.companieshouse.officerfiling.api.enumerations.ValidationEnum;
 import uk.gov.companieshouse.officerfiling.api.error.ApiErrors;
-import uk.gov.companieshouse.officerfiling.api.error.ErrorType;
-import uk.gov.companieshouse.officerfiling.api.error.LocationType;
-import uk.gov.companieshouse.officerfiling.api.exception.CompanyAppointmentServiceException;
-import uk.gov.companieshouse.officerfiling.api.exception.CompanyProfileServiceException;
-import uk.gov.companieshouse.officerfiling.api.exception.ServiceUnavailableException;
 import uk.gov.companieshouse.officerfiling.api.model.dto.OfficerFilingDto;
-import uk.gov.companieshouse.officerfiling.api.service.CompanyAppointmentService;
 import uk.gov.companieshouse.officerfiling.api.service.CompanyProfileService;
 import uk.gov.companieshouse.officerfiling.api.utils.LogHelper;
 
@@ -27,25 +20,15 @@ import uk.gov.companieshouse.officerfiling.api.utils.LogHelper;
  * Provides all validation that should be carried out when an officer is terminated. Fetches all data necessary to complete
  * the validation and generates a list of errors that can be sent back to the caller.
  */
-public class OfficerAppointmentValidator implements OfficerValidator {
+public class OfficerAppointmentValidator extends OfficerValidator {
 
-    private static final List<String> ALLOWED_COMPANY_TYPES = List.of("private-unlimited", "ltd", "plc", "private-limited-guarant-nsc-limited-exemption",
-            "private-limited-guarant-nsc", "private-unlimited-nsc", "private-limited-shares-section-30-exemption");
-    private static final List<String> ALLOWED_OFFICER_ROLES = List.of("director", "corporate-director", "nominee-director", "corporate-nominee-director");
-
-    private final CompanyProfileService companyProfileService;
-    private final CompanyAppointmentService companyAppointmentService;
-    private final Logger logger;
-    private final ApiEnumerations apiEnumerations;
+    private Logger logger;
 
     public OfficerAppointmentValidator(final Logger logger,
                                        final CompanyProfileService companyProfileService,
-                                       final CompanyAppointmentService companyAppointmentService,
                                        final ApiEnumerations apiEnumerations) {
+        super(logger, companyProfileService, apiEnumerations);
         this.logger = logger;
-        this.companyProfileService = companyProfileService;
-        this.companyAppointmentService = companyAppointmentService;
-        this.apiEnumerations = apiEnumerations;
     }
 
     /**
@@ -56,6 +39,7 @@ public class OfficerAppointmentValidator implements OfficerValidator {
      * @param passthroughHeader ERIC pass through header for authorisation
      * @return An object containing a list of any validation errors that have been raised
      */
+    @Override
     public ApiErrors validate(HttpServletRequest request, OfficerFilingDto dto, Transaction transaction, String passthroughHeader) {
             logger.debugContext(transaction.getId(), "Beginning officer appointmnet validation", new LogHelper.Builder(transaction)
                 .withRequest(request)
@@ -82,70 +66,27 @@ public class OfficerAppointmentValidator implements OfficerValidator {
         return new ApiErrors(errorList);
     }
 
+    @Override
     public void validateRequiredDtoFields(HttpServletRequest request, List<ApiError> errorList, OfficerFilingDto dto) {
         // check for blank ???????
 
     }
 
+    @Override
     public void validateRequiredTransactionFields(HttpServletRequest request, List<ApiError> errorList, Transaction transaction) {
         if (transaction.getCompanyNumber() == null || transaction.getCompanyNumber().isBlank()) {
             createValidationError(request, errorList, "The company number cannot be null or blank");
         }
     }
 
-    public Optional<CompanyProfileApi> getCompanyProfile(HttpServletRequest request,
-        List<ApiError> errorList, Transaction transaction, String passthroughHeader) {
-        try {
-            return Optional.ofNullable(
-                companyProfileService.getCompanyProfile(transaction.getId(), transaction.getCompanyNumber(),
-                    passthroughHeader));
-        } catch (ServiceUnavailableException e) {
-            createServiceError(request, errorList);
-        } catch (CompanyProfileServiceException e) {
-            createValidationError(request, errorList, apiEnumerations.getValidation(ValidationEnum.CANNOT_FIND_COMPANY));
-        }
-        return Optional.empty();
-    }
-
+    @Override
     public void validateCompanyNotDissolved(HttpServletRequest request, List<ApiError> errorList, CompanyProfileApi companyProfile) {
         if (companyProfile.getCompanyStatus() == null) {
             logger.errorRequest(request, "null data was found in the Company Profile API within the Company Status field");
             return;
         }
         if (Objects.equals(companyProfile.getCompanyStatus(), "dissolved") || companyProfile.getDateOfCessation() != null) {
-            createValidationError(request, errorList, apiEnumerations.getValidation(ValidationEnum.COMPANY_DISSOLVED));
+            createValidationError(request, errorList, getApiEnumerations().getValidation(ValidationEnum.COMPANY_DISSOLVED));
         }
-    }
-
-    public void validateAllowedCompanyType(HttpServletRequest request, List<ApiError> errorList, CompanyProfileApi companyProfile) {
-        if (companyProfile.getType() == null) {
-            logger.errorRequest(request, "null data was found in the Company Profile API within the Company Type field");
-            return;
-        }
-        if (!ALLOWED_COMPANY_TYPES.contains(companyProfile.getType())) {
-            createValidationError(request, errorList, apiEnumerations.getValidation(ValidationEnum.COMPANY_TYPE_NOT_PERMITTED, apiEnumerations.getCompanyType(companyProfile.getType())));
-        }
-    }
-
-    public void validateOfficerRole(HttpServletRequest request, List<ApiError> errorList, AppointmentFullRecordAPI companyAppointment) {
-        if (companyAppointment.getOfficerRole() == null) {
-            logger.errorRequest(request, "null data was found in the Company Appointment API within the Officer Role field");
-            return;
-        }
-        if (!ALLOWED_OFFICER_ROLES.contains(companyAppointment.getOfficerRole())) {
-            createValidationError(request, errorList, apiEnumerations.getValidation(ValidationEnum.OFFICER_ROLE));
-        }
-    }
-
-    public void createServiceError (HttpServletRequest request, List<ApiError> errorList) {
-        final var apiError = new ApiError(apiEnumerations.getValidation(ValidationEnum.SERVICE_UNAVAILABLE), request.getRequestURI(),
-            LocationType.JSON_PATH.getValue(), ErrorType.SERVICE.getType());
-        errorList.add(apiError);
-    }
-
-    public void createValidationError(HttpServletRequest request, List<ApiError> errorList, String errorMessage) {
-        final var apiError = new ApiError(errorMessage, request.getRequestURI(),
-                LocationType.JSON_PATH.getValue(), ErrorType.VALIDATION.getType());
-        errorList.add(apiError);
     }
 }
